# feat: Complete application directory redesign with modern UI/UX