--- a/components/layout/Header/Header.tsx
+++ b/components/layout/Header/Header.tsx
@@ -14,7 +14,7 @@ export default async function Header({ locale }: { locale: string }) {
   }
 
   return (
-    <header className="border-b bg-white shadow-sm sticky top-0 z-50" role="banner">
+    <header className="sticky top-0 z-50 w-full border-b bg-white/95 backdrop-blur supports-[backdrop-filter]:bg-white/60 shadow-sm" role="banner">
       <div className="container mx-auto px-4 md:px-6 lg:px-8 py-4 flex items-center justify-between">
         <Logo locale={locale} size="md" />
         <div className="flex items-center space-x-4">