--- a/components/sections/Features/Features.tsx
+++ b/components/sections/Features/Features.tsx
@@ -16,7 +16,7 @@ interface FeaturesProps {
 
 export default function Features({ locale, description, features }: FeaturesProps) {
   return (
-    <section id="features" className="py-20 px-4 md:px-6 lg:px-8 bg-gray-50">
+    <section id="features" className="py-20 px-4 md:px-6 lg:px-8 bg-gray-50 scroll-mt-20">
       <div className="container mx-auto max-w-6xl">
         <div className="text-center mb-16">
           <h2 className="text-3xl md:text-4xl font-bold mb-4 text-gray-900">