… AGENTS.md
… CLAUDE.md
… README.md
… app/[locale]/page.tsx.backup
… bun.lock
… package.json